remove some test stuff
